requirements.txt
# View API (MongoDB) + Master Data (PostgreSQL / SQLAlchemy).
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv>=1.0.0
pymongo>=4.9.0,<4.10.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
